+++ internal/conn/memory/permission/permission.go
@@ -22,7 +22,8 @@ func init() {
 type backend struct {
 	users map[string]util.Set
 	perms map[riposo.Path]map[string]util.Set
-	mu    sync.RWMutex
+
+	mu sync.Mutex
 }
 
 // New inits a new in-memory permission backend. Please use for development and testing only!
@@ -40,7 +41,8 @@ func (*backend) Ping(_ context.Context) error {
 
 // Begin implements permission.Backend interface.
 func (b *backend) Begin(_ context.Context) (permission.Transaction, error) {
-	return b, nil
+	b.mu.Lock()
+	return &transaction{b: b}, nil
 }
 
 // Close implements permission.Backend interface.
@@ -48,28 +50,82 @@ func (*backend) Close() error {
 	return nil
 }
 
+// --------------------------------------------------------------------
+
+type transaction struct {
+	b *backend
+
+	xusers map[string]util.Set
+	xperms map[riposo.Path]map[string]util.Set
+
+	done, flushed bool
+}
+
 // Commit implements permission.Transaction interface.
-func (*backend) Commit() error { return nil }
+func (t *transaction) Commit() error {
+	if t.done {
+		return permission.ErrTxDone
+	}
+	t.done = true
+	defer t.b.mu.Unlock()
+
+	return nil
+}
 
 // Rollback implements permission.Transaction interface.
-func (*backend) Rollback() error { return nil }
+func (t *transaction) Rollback() error {
+	if t.done {
+		return permission.ErrTxDone
+	}
+	t.done = true
+	defer t.b.mu.Unlock()
+
+	if t.flushed {
+		t.b.users, t.b.perms = t.xusers, t.xperms
+		return nil
+	}
+
+	for k, v := range t.xusers {
+		if v == nil {
+			delete(t.b.users, k)
+		} else {
+			t.b.users[k] = v
+		}
+	}
+	for k, v := range t.xperms {
+		if v == nil {
+			delete(t.b.perms, k)
+		} else {
+			t.b.perms[k] = v
+		}
+	}
+
+	return nil
+}
 
 // Flush implements permission.Transaction interface.
-func (b *backend) Flush() error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
+func (t *transaction) Flush() error {
+	if t.done {
+		return permission.ErrTxDone
+	}
 
-	b.users = make(map[string]util.Set)
-	b.perms = make(map[riposo.Path]map[string]util.Set)
+	if !t.flushed {
+		t.flushed = true
+		t.xusers = t.b.users
+		t.xperms = t.b.perms
+	}
+	t.b.users = make(map[string]util.Set)
+	t.b.perms = make(map[riposo.Path]map[string]util.Set)
 	return nil
 }
 
 // GetUserPrincipals implements permission.Transaction interface.
-func (b *backend) GetUserPrincipals(userID string) ([]string, error) {
-	b.mu.RLock()
-	defer b.mu.RUnlock()
+func (t *transaction) GetUserPrincipals(userID string) ([]string, error) {
+	if t.done {
+		return nil, permission.ErrTxDone
+	}
 
-	set := b.users[userID].Copy()
+	set := t.b.users[userID].Copy()
 	set.Add(userID)
 
 	switch userID {
@@ -77,26 +133,29 @@ func (b *backend) GetUserPrincipals(userID string) ([]string, error) {
 		// pass-through
 	case riposo.Authenticated:
 		set.Add(riposo.Everyone)
-		set.Merge(b.users[riposo.Everyone])
+		set.Merge(t.b.users[riposo.Everyone])
 	default:
 		set.Add(riposo.Authenticated)
 		set.Add(riposo.Everyone)
-		set.Merge(b.users[riposo.Authenticated])
-		set.Merge(b.users[riposo.Everyone])
+		set.Merge(t.b.users[riposo.Authenticated])
+		set.Merge(t.b.users[riposo.Everyone])
 	}
 	return set.Slice(), nil
 }
 
 // AddUserPrincipal implements permission.Transaction interface.
-func (b *backend) AddUserPrincipal(principal string, userIDs []string) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
+func (t *transaction) AddUserPrincipal(principal string, userIDs []string) error {
+	if t.done {
+		return permission.ErrTxDone
+	}
 
 	for _, userID := range userIDs {
-		set, ok := b.users[userID]
+		t.backupUser(userID)
+
+		set, ok := t.b.users[userID]
 		if !ok {
 			set := util.NewSet(principal)
-			b.users[userID] = set
+			t.b.users[userID] = set
 		} else {
 			set.Add(principal)
 		}
@@ -105,12 +164,15 @@ func (b *backend) AddUserPrincipal(principal string, userIDs []string) error {
 }
 
 // RemoveUserPrincipal implements permission.Transaction interface.
-func (b *backend) RemoveUserPrincipal(principal string, userIDs []string) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
+func (t *transaction) RemoveUserPrincipal(principal string, userIDs []string) error {
+	if t.done {
+		return permission.ErrTxDone
+	}
 
 	for _, userID := range userIDs {
-		if set, ok := b.users[userID]; ok {
+		if set, ok := t.b.users[userID]; ok {
+			t.backupUser(userID)
+
 			set.Remove(principal)
 		}
 	}
@@ -118,12 +180,15 @@ func (b *backend) RemoveUserPrincipal(principal string, userIDs []string) error 
 }
 
 // PurgeUserPrincipals implements permission.Transaction interface.
-func (b *backend) PurgeUserPrincipals(principals []string) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
+func (t *transaction) PurgeUserPrincipals(principals []string) error {
+	if t.done {
+		return permission.ErrTxDone
+	}
 
 	for _, principal := range principals {
-		for _, set := range b.users {
+		for userID, set := range t.b.users {
+			t.backupUser(userID)
+
 			set.Remove(principal)
 		}
 	}
@@ -131,11 +196,12 @@ func (b *backend) PurgeUserPrincipals(principals []string) error {
 }
 
 // GetACEPrincipals implements permission.Transaction interface.
-func (b *backend) GetACEPrincipals(ent permission.ACE) ([]string, error) {
-	b.mu.RLock()
-	defer b.mu.RUnlock()
+func (t *transaction) GetACEPrincipals(ent permission.ACE) ([]string, error) {
+	if t.done {
+		return nil, permission.ErrTxDone
+	}
 
-	if perms, ok := b.perms[ent.Path]; ok {
+	if perms, ok := t.b.perms[ent.Path]; ok {
 		if set, ok := perms[ent.Perm]; ok {
 			return set.Slice(), nil
 		}
@@ -144,14 +210,17 @@ func (b *backend) GetACEPrincipals(ent permission.ACE) ([]string, error) {
 }
 
 // AddACEPrincipal implements permission.Transaction interface.
-func (b *backend) AddACEPrincipal(principal string, ent permission.ACE) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
+func (t *transaction) AddACEPrincipal(principal string, ent permission.ACE) error {
+	if t.done {
+		return permission.ErrTxDone
+	}
 
-	perms, ok := b.perms[ent.Path]
+	t.backupPerms(ent.Path)
+
+	perms, ok := t.b.perms[ent.Path]
 	if !ok {
 		perms = make(map[string]util.Set)
-		b.perms[ent.Path] = perms
+		t.b.perms[ent.Path] = perms
 	}
 
 	set, ok := perms[ent.Perm]
@@ -164,11 +233,14 @@ func (b *backend) AddACEPrincipal(principal string, ent permission.ACE) error {
 }
 
 // RemoveACEPrincipal implements permission.Transaction interface.
-func (b *backend) RemoveACEPrincipal(principal string, ent permission.ACE) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
+func (t *transaction) RemoveACEPrincipal(principal string, ent permission.ACE) error {
+	if t.done {
+		return permission.ErrTxDone
+	}
 
-	if perms, ok := b.perms[ent.Path]; ok {
+	if perms, ok := t.b.perms[ent.Path]; ok {
+		t.backupPerms(ent.Path)
+
 		if set, ok := perms[ent.Perm]; ok {
 			set.Remove(principal)
 		}
@@ -177,18 +249,18 @@ func (b *backend) RemoveACEPrincipal(principal string, ent permission.ACE) error
 }
 
 // GetAllACEPrincipals implements permission.Transaction interface.
-func (b *backend) GetAllACEPrincipals(ents []permission.ACE) ([]string, error) {
+func (t *transaction) GetAllACEPrincipals(ents []permission.ACE) ([]string, error) {
+	if t.done {
+		return nil, permission.ErrTxDone
+	}
+
 	if len(ents) == 0 {
 		return nil, nil
 	}
 
 	res := util.NewSet()
-
-	b.mu.RLock()
-	defer b.mu.RUnlock()
-
 	for _, ent := range ents {
-		if perms, ok := b.perms[ent.Path]; ok {
+		if perms, ok := t.b.perms[ent.Path]; ok {
 			if set, ok := perms[ent.Perm]; ok {
 				res.Merge(set)
 			}
@@ -198,26 +270,30 @@ func (b *backend) GetAllACEPrincipals(ents []permission.ACE) ([]string, error) {
 }
 
 // GetPermissions implements permission.Transaction interface.
-func (b *backend) GetPermissions(path riposo.Path) (schema.PermissionSet, error) {
-	b.mu.RLock()
-	defer b.mu.RUnlock()
+func (t *transaction) GetPermissions(path riposo.Path) (schema.PermissionSet, error) {
+	if t.done {
+		return nil, permission.ErrTxDone
+	}
 
-	perms := make(schema.PermissionSet, len(b.perms[path]))
-	for perm, set := range b.perms[path] {
+	perms := make(schema.PermissionSet, len(t.b.perms[path]))
+	for perm, set := range t.b.perms[path] {
 		perms[perm] = set.Slice()
 	}
 	return perms, nil
 }
 
 // CreatePermissions implements permission.Transaction interface.
-func (b *backend) CreatePermissions(path riposo.Path, set schema.PermissionSet) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
+func (t *transaction) CreatePermissions(path riposo.Path, set schema.PermissionSet) error {
+	if t.done {
+		return permission.ErrTxDone
+	}
 
-	perms, ok := b.perms[path]
+	t.backupPerms(path)
+
+	perms, ok := t.b.perms[path]
 	if !ok {
 		perms = make(map[string]util.Set, len(set))
-		b.perms[path] = perms
+		t.b.perms[path] = perms
 	}
 
 	for perm, principals := range set {
@@ -234,14 +310,17 @@ func (b *backend) CreatePermissions(path riposo.Path, set schema.PermissionSet) 
 }
 
 // MergePermissions implements permission.Transaction interface.
-func (b *backend) MergePermissions(path riposo.Path, set schema.PermissionSet) error {
-	b.mu.Lock()
-	defer b.mu.Unlock()
+func (t *transaction) MergePermissions(path riposo.Path, set schema.PermissionSet) error {
+	if t.done {
+		return permission.ErrTxDone
+	}
 
-	perms, ok := b.perms[path]
+	t.backupPerms(path)
+
+	perms, ok := t.b.perms[path]
 	if !ok {
 		perms = make(map[string]util.Set, len(set))
-		b.perms[path] = perms
+		t.b.perms[path] = perms
 	}
 	for perm, principals := range set {
 		if len(principals) == 0 {
@@ -254,18 +333,20 @@ func (b *backend) MergePermissions(path riposo.Path, set schema.PermissionSet) e
 }
 
 // DeletePermissions implements permission.Transaction interface.
-func (b *backend) DeletePermissions(paths []riposo.Path) error {
+func (t *transaction) DeletePermissions(paths []riposo.Path) error {
+	if t.done {
+		return permission.ErrTxDone
+	}
+
 	if len(paths) == 0 {
 		return nil
 	}
 
-	b.mu.Lock()
-	defer b.mu.Unlock()
-
-	for path := range b.perms {
+	for path := range t.b.perms {
 		for _, pattern := range paths {
 			if pattern == path || strings.HasPrefix(path.String(), pattern.String()+"/") {
-				delete(b.perms, path)
+				t.backupPerms(path)
+				delete(t.b.perms, path)
 			}
 		}
 	}
@@ -273,15 +354,16 @@ func (b *backend) DeletePermissions(paths []riposo.Path) error {
 }
 
 // GetAccessiblePaths implements permission.Transaction interface.
-func (b *backend) GetAccessiblePaths(dst []riposo.Path, principals []string, ents []permission.ACE) ([]riposo.Path, error) {
+func (t *transaction) GetAccessiblePaths(dst []riposo.Path, principals []string, ents []permission.ACE) ([]riposo.Path, error) {
+	if t.done {
+		return nil, permission.ErrTxDone
+	}
+
 	if len(principals) == 0 || len(ents) == 0 {
 		return nil, nil
 	}
 
-	b.mu.RLock()
-	defer b.mu.RUnlock()
-
-	for path, perms := range b.perms {
+	for path, perms := range t.b.perms {
 		for perm, allowed := range perms {
 			if match(ents, path, perm) && allowed.HasAny(principals...) {
 				dst = append(dst, path)
@@ -292,6 +374,38 @@ func (b *backend) GetAccessiblePaths(dst []riposo.Path, principals []string, ent
 
 	sort.Slice(dst, func(i, j int) bool { return dst[i] < dst[j] })
 	return dst, nil
+}
+
+func (t *transaction) backupUser(userID string) {
+	if t.flushed {
+		return
+	}
+
+	if t.xusers == nil {
+		t.xusers = make(map[string]util.Set)
+	}
+
+	if _, ok := t.xusers[userID]; !ok {
+		t.xusers[userID] = t.b.users[userID].Copy()
+	}
+}
+
+func (t *transaction) backupPerms(path riposo.Path) {
+	if t.flushed {
+		return
+	}
+
+	if t.xperms == nil {
+		t.xperms = make(map[riposo.Path]map[string]util.Set)
+	}
+
+	if _, ok := t.xperms[path]; !ok {
+		pristine := t.b.perms[path]
+		t.xperms[path] = make(map[string]util.Set, len(pristine))
+		for k, v := range pristine {
+			t.xperms[path][k] = v.Copy()
+		}
+	}
 }
 
 func match(ents []permission.ACE, path riposo.Path, perm string) bool {