+++ internal/conn/memory/storage/storage_test.go
@@ -32,7 +32,9 @@ var _ = Describe("Backend", func() {
 		Expect(subject.Close()).To(Succeed())
 	})
 
-	testdata.BehavesLikeBackend(&link)
+	Describe("common", func() {
+		testdata.BehavesLikeBackend(&link)
+	})
 })
 
 func TestSuite(t *testing.T) {